fix bug for gnu compiler

+++ source/src_pw/bspline_sf.cpp
@@ -117,26 +117,26 @@ void PW_Basis:: bsplinecoef(complex<double> *b1, complex<double> *b2, complex<do
         complex<double> fracx=0;
         for(int io = 0 ; io < norder - 1 ; ++io)
         {
-            fracx += bsp.bezier_ele(io)*exp(ci_tpi*ix/double(nx)*io);
+            fracx += bsp.bezier_ele(io)*exp(ci_tpi*double(ix)/double(nx)*io);
         }
-        b1[ix] = exp(ci_tpi*norder*ix/nx)/fracx;
+        b1[ix] = exp(ci_tpi*double(norder*ix)/double(nx))/fracx;
     }
     for(int iy = 0 ; iy < this->ny ; ++iy)
     {
         complex<double> fracy=0;
         for(int io = 0 ; io < norder - 1 ; ++io)
         {
-            fracy += bsp.bezier_ele(io)*exp(ci_tpi*iy/double(ny)*io);
+            fracy += bsp.bezier_ele(io)*exp(ci_tpi*double(iy)/double(ny)*io);
         }
-        b2[iy] = exp(ci_tpi*norder*iy/ny)/fracy;
+        b2[iy] = exp(ci_tpi*double(norder*iy)/double(ny))/fracy;
     }
     for(int iz = 0 ; iz < this->nz ; ++iz)
     {
         complex<double> fracz=0;
         for(int io = 0 ; io < norder - 1 ; ++io)
         {
-            fracz += bsp.bezier_ele(io)*exp(ci_tpi*iz/double(nz)*io);
+            fracz += bsp.bezier_ele(io)*exp(ci_tpi*double(iz)/double(nz)*io);
         }
-        b3[iz] = exp(ci_tpi*norder*iz/nz)/fracz;
+        b3[iz] = exp(ci_tpi*double(norder*iz)/double(nz))/fracz;
     }
 }
